+++ src/RcsbChartImplementations/ChatJsImplementations/Components/BarComponent.ts
@@ -1,11 +1,11 @@
 import Chart, {ActiveElement, ChartEvent} from "chart.js/auto";
 import {DataContainerReader} from "../../../Utils/DataContainer";
-import {ChartDataInterface} from "../../../RcsbChartDataProvider/ChartDataProviderInterface";
+import {ChartDataColumnInterface} from "../../../RcsbChartDataProvider/ChartDataProviderInterface";
 import {BarClickCallbackType} from "../../../RcsbChartComponent/ChartConfigInterface";
 import Element from "chart.js/dist/core/core.element";
 
-type RawType = {x:string|number; y:number; id:any;};
-export function chartJsBarClick(dataContainer: DataContainerReader<ChartDataInterface[]>, barClickCallback?:BarClickCallbackType) {
+type RawType = {x:string|number; y:number; id:unknown;};
+export function chartJsBarClick(dataContainer: DataContainerReader<ChartDataColumnInterface[]>, barClickCallback?:BarClickCallbackType) {
     if(! barClickCallback)
         return undefined;
     return (event: ChartEvent, elements: ActiveElement[], chart: Chart) => {
@@ -13,7 +13,8 @@ export function chartJsBarClick(dataContainer: DataContainerReader<ChartDataInte
         if(!element)
             return;
         barClickCallback({
-            y: dataContainer.get()?.[element.$context.parsed.y].y.filter(d=>d.value>0) ?? [],
+            values: dataContainer.get()?.[element.$context.parsed.y].y.filter(d=>d.value>0) ?? [],
+            y: element.$context.raw.y,
             x: element.$context.raw.x,
             id: element.$context.raw.id
         }, dataContainer.get() ?? []);