+++ src/RcsbChartImplementations/ChatJsImplementations/Components/TootlipComponent.ts
@@ -1,8 +1,10 @@
-import {ChartDataValuesInterface} from "../../../RcsbChartDataProvider/ChartDataProviderInterface";
+import {
+    ChartDataValueInterface
+} from "../../../RcsbChartDataProvider/ChartDataProviderInterface";
 import {TooltipItem} from "chart.js/auto";
 
 type RawType = {x:string|number; y:number; id:any;};
-export function chartJsTooltip(tooltipText?:(a: ChartDataValuesInterface) => string|undefined) {
+export function chartJsTooltip(tooltipText?:(a: ChartDataValueInterface) => string|undefined) {
     if(!tooltipText)
         return undefined;
     return {
@@ -10,9 +12,8 @@ export function chartJsTooltip(tooltipText?:(a: ChartDataValuesInterface) => str
             afterLabel: (tooltipItem: TooltipItem<any>) =>{
                 const raw: RawType = tooltipItem.raw as RawType;
                 return tooltipText?.({
-                    y: raw.y,
                     values: tooltipItem.dataset.data,
-                    index: tooltipItem.datasetIndex,
+                    y: raw.y,
                     x: raw.x,
                     id: raw.id
                 })