+++ src/RcsbChartImplementations/VictoryChartImplementations/VictoryBarChartComponent.tsx
@@ -6,12 +6,13 @@ import {LabelComponent} from "./Components/LabelComponent";
 import * as React from "react";
 import {BarComponent} from "./Components/BarComponent";
 import {TooltipFactory} from "./Components/TooltipFactory";
-import {ChartDataInterface, ChartDataValuesInterface} from "../../RcsbChartDataProvider/ChartDataProviderInterface";
+import {ChartDataColumnInterface, ChartDataValueInterface} from "../../RcsbChartDataProvider/ChartDataProviderInterface";
 import {ChartConfigInterface, ChartDisplayConfigInterface} from "../../RcsbChartComponent/ChartConfigInterface";
+import {VictoryChartDataInterface} from "./VictoryChartDataInterface";
 
 export class VictoryBarChartComponent extends AbstractChartImplementation {
     render():JSX.Element {
-        const {data}: {data: ChartDataInterface[]; excludedData?:ChartDataInterface[];} = this.props.dataProvider.getChartData();
+        const {data}: {data: ChartDataColumnInterface[]; excludedData?:ChartDataColumnInterface[];} = this.props.dataProvider.getChartData();
         const displayConfig: Partial<ChartDisplayConfigInterface> = this.props.chartConfig?.chartDisplayConfig ?? {};
         return (data.length == 0 ? <></> : <VictoryChart
             domainPadding={{ x: ChartTools.getConfig<number>("xDomainPadding",displayConfig) }}
@@ -28,7 +29,7 @@ export class VictoryBarChartComponent extends AbstractChartImplementation {
 }
 
 //TODO <VictoryStack animate={true}> BarComponent props fails in capturing updated data
-function stack(data:ChartDataInterface[], chartConfig?: ChartConfigInterface): JSX.Element{
+function stack(data:ChartDataColumnInterface[], chartConfig?: ChartConfigInterface): JSX.Element{
     return ( <VictoryStack >
         {bar(
             data.map(d=>({
@@ -36,7 +37,7 @@ function stack(data:ChartDataInterface[], chartConfig?: ChartConfigInterface): J
                 y:d.y[0].value,
                 color:d.y[0].color,
                 id:d.y[0].id,
-                values:d.y.map(v=>v.value),
+                values:d.y,
                 index:0
             })),
             0,
@@ -54,7 +55,7 @@ function stack(data:ChartDataInterface[], chartConfig?: ChartConfigInterface): J
                             y:d.y[n+1].value,
                             color:d.y[n+1].color,
                             id:d.y[n+1].id,
-                            values:d.y.map(v=>v.value),
+                            values: d.y,
                             index:n+1
                         })),
                         n+1,
@@ -70,7 +71,7 @@ function stack(data:ChartDataInterface[], chartConfig?: ChartConfigInterface): J
     </VictoryStack>);
 }
 
-function bar(data:ChartDataValuesInterface[], index:number, color: string, barComp: JSX.Element, labelComponent?:JSX.Element, chartDisplayConfig?:Partial<ChartDisplayConfigInterface>): JSX.Element {
+function bar(data:VictoryChartDataInterface[], index:number, color: string, barComp: JSX.Element, labelComponent?:JSX.Element, chartDisplayConfig?:Partial<ChartDisplayConfigInterface>): JSX.Element {
     return data.length > 0 ? (<VictoryBar
         key={"victory_bar_"+index}
         barWidth={ChartTools.getConfig<number>("xDomainPadding", chartDisplayConfig)}