+++ src/RcsbChartImplementations/VictoryChartImplementations/VictoryChartDataInterface.ts
@@ -1,0 +1,10 @@
+
+export interface VictoryChartDataInterface {
+    x:string|number;
+    y:number;
+    values:{value:number; color?:string; id?:any;}[];
+    index:number;
+    label?:string;
+    color?:string;
+    id?:any;
+}